kidverse/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
jinja2==3.1.4
python-multipart==0.0.18
httpx==0.28.1
aiofiles==24.1.0
python-dotenv==1.0.1
bcrypt==4.2.1
pyjwt==2.10.1
